Move monitoring types according to new structure
